src/requirements.txt
colorama
mysql-connector-python